+++ types/index.d.ts
@@ -8,23 +8,9 @@ export interface Coordinates {
     readonly latitude: number | string;
     readonly longitude: number | string;
 }
-<<<<<<< HEAD
-export * from "providers";
-export * from "query";
-export { default as AdminLevel } from "AdminLevel";
-export * from "AdminLevel";
-export { default as ExternalURILoader } from "ExternalURILoader";
-export * from "ExternalURILoader";
-export { default as Geocoded } from "Geocoded";
-export * from "Geocoded";
-export { default as GeocoderProviderFactory } from "GeocoderProviderFactory";
-export * from "GeocoderProviderFactory";
-export { default as GeoJsonDumper } from "GeoJsonDumper";
-export * from "GeoJsonDumper";
-export { default } from "GeocoderJS";
-=======
 export * from "./provider";
 export * from "./query";
+export * from "./utils";
 export { default as AdminLevel } from "./AdminLevel";
 export * from "./AdminLevel";
 export { default as ExternalLoader } from "./ExternalLoader";
@@ -35,6 +21,5 @@ export { default as GeocoderProviderFactory } from "./GeocoderProviderFactory";
 export * from "./GeocoderProviderFactory";
 export { default as GeoJsonDumper } from "./GeoJsonDumper";
 export * from "./GeoJsonDumper";
-export { default } from "./UniversalGeocoder";
->>>>>>> main
+export { default } from "./GeocoderJS";
 //# sourceMappingURL=index.d.ts.map