+++ webpack.common.js
@@ -21,4 +21,5 @@ module.exports = {
   output: {
     path: path.resolve(__dirname, "dist"),
   },
+  node: false,
 };