+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## O.3.0
+## 0.4.0
+
+* [Google] Better Google Maps provider
+* [Nominatim] Better handling of geodecode errors
+* [ExternalLoader] Add `getOptions` method to interface
+* [ExternalLoader] Rename `JSONPCallback` in params to `jsonpCallback`
+* Rename `GoogleAPIProvider` to `GoogleMapsProvider`
+* Add Node examples
+
+## 0.3.0
 
 * Add OpenCage Provider
 * Add time zone to `Geocoded`